requirements.txt
streamlit
groq
folium
streamlit-folium
speechrecognition
requests
google-generativeai
Pillow